refactor: use PeerAddress when connecting to Peers and in events
As `Peer`s are modeled now, it seems most sensible for the DB logic to
be responsible for creating new `Peer` records. The rest of the
application has no extra need for anything but the `PeerAddress` of the
peer, so they can restrict their arguments as such, which also
simplifies publishing and consuming peer events.